Move getFirstBlockIdOfLastRounds() from entity Block to custom methods
